Add VAE float16 option.

+++ src/shearest/data_gen_utils.py
@@ -167,7 +167,8 @@ def image_to_kimage(image, pixel_scale, Npx_out, pixel_scale_out):
 
 def draw_NN_profile(z, flux, g1, g2, key, uv_pos, Npx, pixel_scale_radio, pixel_scale_vae=0.03, jitted_decode=None, gsparams=None, use_dropout=False):
     # Decode the latent vector to get the galaxy image
-    y = jitted_decode(z[None,:,:], key=key if use_dropout else None)[0]  # (Ny, Nx)
+    # Cast to float32 to ensure shear+FFT pipeline runs in full precision
+    y = jitted_decode(z[None,:,:], key=key if use_dropout else None)[0].astype(jnp.float32)
 
     # Normalize to desired flux: total flux = sum(I) * pixel_scale^2
     y = y * flux / (jnp.sum(y) * pixel_scale_vae**2)
